Move DJList components into folder
