+++ client/react/frontpage/components/DJList/DJList.scss
@@ -4,7 +4,7 @@ DJList.scss
 Sass styling for frontpage djs list
 **/
 
-@import "../common";
+@import "../../common";
 
 // constants
 $overlay-opacity: 0.8;
